+++ lib/rspec/rails/example/rails_example_group.rb
@@ -18,6 +18,10 @@ module RSpec
       include RSpec::Rails::TaggedLoggingAdapter
       include ActiveSupport::CurrentAttributes::TestHelper
       include ActiveSupport::ExecutionContext::TestHelper
+
+      if RSpec::Rails::FeatureCheck.has_active_job?
+        include ActiveJob::TestHelper
+      end
     end
   end
 end